+++ packages/nextra-theme-docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # nextra-theme-docs
 
+## 3.1.0-canary.0
+
+### Minor Changes
+
+- fec399a: fix `type: 'separator'`, `type: 'menu'` and `item` with `href` not
+  respecting order when not all pages specified in `_meta` file
+
+### Patch Changes
+
+- Updated dependencies [fec399a]
+  - nextra@3.1.0-canary.0
+
 ## 3.0.15
 
 ### Patch Changes